Extract waitFor functions to separated file
