+++ test/utils/client.dart
@@ -6,7 +6,6 @@ import 'package:crypto/crypto.dart';
 import 'package:meilisearch/src/http_request_impl.dart';
 import 'package:meilisearch/src/http_request.dart';
 import 'package:meilisearch/meilisearch.dart';
-import 'package:meilisearch/src/task.dart';
 import 'package:test/test.dart';
 
 late HttpRequest http;
@@ -35,8 +34,6 @@ Future<void> setUpClient() async {
   setUp(() {
     final String server = testServer;
 
-    print('Using Meilisearch server on $server for running tests.');
-
     client = MeiliSearchClient(server, 'masterKey');
     random = Random();
   });
@@ -60,9 +57,6 @@ Future<void> setUpClientWithWrongUrl() async {
     final String server = 'http://wrongurl:1234';
     final int connectTimeout = 1000;
 
-    print(
-        'Using wrong url server on $server with timeout of $connectTimeout ms running tests.');
-
     client = MeiliSearchClient(server, 'masterKey', connectTimeout);
   });
 }
@@ -77,37 +71,4 @@ String sha1RandomString() {
   final randomBytes = utf8.encode(randomNumber.toString());
 
   return sha1.convert(randomBytes).toString();
-}
-
-extension TaskWaiter on Task {
-  Future<Task> waitFor({
-    Duration timeout = const Duration(seconds: 5),
-    Duration interval = const Duration(milliseconds: 50),
-  }) async {
-    var endingTime = DateTime.now().add(timeout);
-
-    while (DateTime.now().isBefore(endingTime)) {
-      var task = await client.getTask(this.uid!);
-
-      if (task.status != 'enqueued' && task.status != 'processing') {
-        return task;
-      }
-
-      await Future.delayed(interval);
-    }
-
-    throw Exception('The task ${this.uid!} timed out.');
-  }
-}
-
-extension TaskWaiterForFutures on Future<Task> {
-  Future<Task> waitFor({
-    Duration timeout = const Duration(seconds: 5),
-    Duration interval = const Duration(milliseconds: 50),
-  }) async {
-    return await (await this).waitFor(
-      timeout: timeout,
-      interval: interval,
-    );
-  }
 }
